1132 refactor code with platform-mail module
