+++ server/libs/config/tenant-single-security-config/src/test/java/com/bytechef/tenant/single/security/SingleTenantUserDetailsServiceIntTest.java
@@ -22,9 +22,9 @@ import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import com.bytechef.config.ApplicationProperties;
 import com.bytechef.jdbc.config.AuditingJdbcConfiguration;
 import com.bytechef.liquibase.config.LiquibaseConfiguration;
+import com.bytechef.platform.mail.MailService;
 import com.bytechef.platform.security.exception.UserNotActivatedException;
 import com.bytechef.platform.user.domain.User;
-import com.bytechef.platform.user.service.MailService;
 import com.bytechef.platform.user.service.UserService;
 import com.bytechef.security.config.SecurityConfiguration;
 import com.bytechef.test.config.testcontainers.PostgreSQLContainerConfiguration;
