+++ server/libs/platform/platform-configuration/platform-configuration-service/src/main/java/com/bytechef/platform/configuration/notification/EmailNotificationSender.java
@@ -17,7 +17,7 @@
 package com.bytechef.platform.configuration.notification;
 
 import com.bytechef.platform.configuration.domain.notification.Notification;
-import com.bytechef.platform.user.service.MailService;
+import com.bytechef.platform.mail.MailService;
 import java.util.Map;
 import org.springframework.stereotype.Component;
