+++ server/libs/platform/platform-configuration/platform-configuration-service/src/test/java/com/bytechef/platform/configuration/service/NotificationServiceIntTest.java
@@ -23,23 +23,18 @@ import com.bytechef.platform.configuration.domain.notification.Event;
 import com.bytechef.platform.configuration.domain.notification.Notification;
 import com.bytechef.platform.configuration.repository.EventRepository;
 import com.bytechef.platform.configuration.repository.NotificationRepository;
-import com.bytechef.platform.user.service.MailService;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
 /**
  * @author Matija Petanjek
  */
 @SpringBootTest(classes = NotificationIntTestConfiguration.class)
 public class NotificationServiceIntTest {
-
-    @MockitoBean
-    private MailService mailService;
 
     @Autowired
     private EventRepository eventRepository;
