+++ server/libs/platform/platform-user/platform-user-rest/platform-user-rest-impl/build.gradle.kts
@@ -50,6 +50,7 @@ dependencies {
     implementation(project(":server:libs:core:tenant:tenant-api"))
     implementation(project(":server:libs:core:rest:rest-api"))
     implementation(project(":server:libs:platform:platform-api"))
+    implementation(project(":server:libs:platform:platform-mail"))
     implementation(project(":server:libs:platform:platform-user:platform-user-api"))
     implementation(project(":server:libs:platform:platform-user:platform-user-rest:platform-user-rest-api"))
 
@@ -69,6 +70,7 @@ dependencies {
     testImplementation(project(":server:libs:config:security-config"))
     testImplementation(project(":server:libs:core:tenant:tenant-single-service"))
     testImplementation(project(":server:libs:core:rest:rest-impl"))
+    testImplementation(project(":server:libs:platform:platform-mail"))
     testImplementation(project(":server:libs:platform:platform-user:platform-user-service"))
     testImplementation(project(":server:libs:test:test-int-support"))
