+++ server/libs/platform/platform-user/platform-user-rest/platform-user-rest-impl/src/test/java/com/bytechef/platform/user/web/rest/config/UserIntTestConfiguration.java
@@ -35,7 +35,8 @@ import org.springframework.mail.javamail.JavaMailSenderImpl;
  * @author Ivica Cardic
  */
 @ComponentScan(basePackages = {
-    "com.bytechef.platform.user", "com.bytechef.web.rest", "com.bytechef.tenant", "com.bytechef.ee.tenant"
+    "com.bytechef.platform.mail", "com.bytechef.platform.user", "com.bytechef.web.rest", "com.bytechef.tenant",
+    "com.bytechef.ee.tenant"
 })
 @EnableAutoConfiguration
 @EnableConfigurationProperties(ApplicationProperties.class)
