+++ server/libs/platform/platform-user/platform-user-service/src/test/java/com/bytechef/platform/user/config/UserIntTestConfiguration.java
@@ -27,8 +27,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.JavaMailSenderImpl;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
@@ -41,11 +39,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 })
 @Configuration
 public class UserIntTestConfiguration {
-
-    @Bean
-    JavaMailSender javaMailSender() {
-        return new JavaMailSenderImpl();
-    }
 
     @Bean
     PasswordEncoder passwordEncoder() {
